Fix short test runs for several examples
This fixes examples where passing `short` to `run` didn't have an effect
because the number of iterations was hardcoded in the `sample` call.

The length of the short run in GHME_2013.py is increased so that slicing
the trace object to burn the first 100 iterations does not lead to an
error when it is passed to `traceplot`.