#626: ImportError when IPython not installed
When importing pymc, "interactive_sampling.py" would try to import
IPython. If IPython is not present, importing pymc failed with an import
error.

The solution is to check for the IPython module to be present. This
commit also checks if it is running inside an IPython Notebook kernel
and raises an NotImplementedError otherwise.