Add testval to HalfCauchy in GLM and Fix GLM tutorials.
The HalfCauchy defaults to testval=0 which is not a valid
starting point if used as a parameter for SD.

GLM now also does not initialize by default which is fixed
in the GLM NBs.

Both of those issues were brought up in https://github.com/pymc-devs/pymc3/issues/702.

Moreover, I moved those to the examples directory.

There was also a missing radon.csv data file which I added.