split observedrv into observedrv and multiobserved
With the introduction of missing variables it becomes important to be
able to use the value of an ObservedRV. However, this was not possible
because ObservedRV didn't subclass TensorVariable because ObservedRV
could also handle multiple observed arrays so wouldn't always be a
single tensor.

This splits up the two cases into ObservedRV, for single observed
arrays, and MultiObservedRV for multiple observed arrays. ObservedRV
now inherits from TensorVariable.