Change source of variable names in backend tests
The variables names from the setup trace and test_point should be the
same, but it's safer to use the names from the more upstream
test_point.

Also, fix some places where the variable name was hard-coded to be 'x'
so that the tests don't rely on the specific model used.