Replace Text backend with CSV
The main reason to keep the Text backend around was that pandas was an
optional dependency.  Now that this is no longer the case, the
original Text backend doesn't offer any advantages over the CSV
backend.  (Even if someone prefers not to write to files while
sampling, they can sample with the NDArray backend and then use the
CSV dump function.)

Rename CSV to Text.  This name is more appropriate because the values
being stored as plain text is the important feature, not which
delimiter is used.