Correction to Nomenclature in Volatility Model (#2978)
This small correction is in response to issue Stochastic Volatility Example #2566.  It concerns the
two different examples of the Volatility Model used in the PyMC3 notebooks to introduce users to the
wonder that is Bayesian modeling.  In getting_started.ipynb, under:
--Case Study 1: Stochastic volatility
---The Model
the model specification uses `y_i` to represent the daily precentage returns.  However, later on in
the `with pm.Model() as sp500_model:` block the dummy variable `r` is used to represent the daily
returns as well as the tensor variable name.

The second correction also concerns the useage of `s_i` to represent the volatility process in daily
returns.  In both Volatility Process walkthroughs (there is a stand-alone notebook dedicated to it)
the model specs treat `s_i` as the standard deviation of the StudentT distribution used to model the
log-returns.  In the PyMC3 API docs on the StudentT, the distribution is defined with `lambda`
representing the precision.  This is why the `volatility_process` variable is mapped from
`pm.math.exp(-2 * s)` in the `pm.Model()` block.  However, when the returns were defined in the
model block the **kwarg was `lambda=1/volatility_process`.  This has been fixed.

Thanks to @twiecki and the OP for highlighting this error.