Removed or edited for unsupported features.
The new version of `sample_posterior_predictive()` does not support the `size` argument, so tests using it are commented out.
Also, we don't support asking for a number of samples that is not a round multiple of the number of chains, so in some cases I bumped the number of samples to fit the input trace object.