Replace MultiTrace with new _TraceDict.
Created a new _TraceDict class that represents a trace as a dictionary of variable names to numpy arrays, and use that as the internal representation, per @lucianopaz suggestion.
Made this a class so that it can have varnames and a length, like the MultiTrace.
Added a couple of translation tests.