Get the transform docs into the API pages. (#3690)
* Extend exports from transforms.

* Add docs for distributions.transforms.

The doc strings for the transforms were not linked into the documentation page for distributions.

* Improve documentation of transform aliases.

The transforms code has lower-case alternatives to the upper-cased
class names.  Sometimes these are instances of the classes (e.g.,
`log` is an instance of the `Log` transform), but sometimes they are
aliases (e.g., `lowerbound` is an alias for `LowerBound`).

I couldn't find a good way to document these in the source file, so
they are hand-documented in transforms.rst.

* Export improvements requested by @ColCarroll.

* Minor edits to clarify transformations.

* Added API page for random variables.

* Fix typo @twiecki caught and some others.

Thomas caught a typo of mine, and when fixing it, I picked off some grammar errors and improved some typesetting.

* Add to release notes.

* Fix typesetting of parameters.

Fix typesetting of parameters in the docstrings.

* Add docstring for t_stick_breaking()

Thanks to @aloctavodia for help.

* Slightly improved Table of Contents.

I still don't know how to get the title of *this* page *not* to appear in its table of contents.
Unfortunately, `autosummary` does not work at all.

* Fix formatting of additional parameters.

While I was at it, fixed the punctuation of parameters in the docstrings so that Sphinx will format them correctly.