Extend transforms to non-RandomVariable types
The `TransformedRV` was effectively serving as a means of creating a
`_default_transformed_rv` dispatch and registering the new types created in
`_create_transformed_rv_op` as `MeasurableVariable`s.  This class has been
removed and replaced with a much more direct approach in
`_create_transformed_rv_op`.  If transformed variable types ever really need to
be identified by type alone, we can create another abstract class and register
them in `_create_transformed_rv_op`, just as we do with `MeasurableVariable`.